refactor(sandbox-fc): split create transaction tests
